Allow to add/remove associations to linked Xreport users
